% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{draw_key_sector}
\alias{draw_key_sector}
\title{draw_key_sector}
\usage{
draw_key_sector(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
ggplot legend
}
\description{
draw_key_sector
}
