% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sector_ggplot.R
\docType{data}
\name{GeomSectorPanel}
\alias{GeomSectorPanel}
\alias{GeomSectorIndividual}
\alias{geom_sector}
\title{ggplot sector}
\usage{
geom_sector(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  individual = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{individual}{Logical, default is TRUE.
When "individual=FALSE", draw very quickly with a vector form,
when "individual=TRUE", draw individually at a slower speed.
Vector form cannot control the deformation of the sector, you need to add coord_fixed()}

\item{verbose}{Logical, default is TRUE. Whether to display reminder information.}
}
\value{
ggplot object
}
\description{
Draw sector with ggplot2.
}
\details{
When "individual=FALSE", draw very quickly with a vector form,
when "individual=TRUE", draw individually at a slower speed.

The required parameters in mapping are "x", "y", "theta", and
the additional modifiable parameters are "r", "start", "r_start", "type", "colour", "fill".

When r = 0.5, the sector just fills the square.

For details, please check the \code{\link[=grid.sector]{grid.sector()}}.

For more details, please type \code{vignette("ggsector")}.
}
\examples{
\donttest{
# prepare
library(ggsector)
library(reshape2)
df <- cor(mtcars)[1:3, 1:5] \%>\%
    abs() \%>\%
    melt(varnames = c("x", "y"))

# individual
ggplot(df) +
    geom_sector(aes(x, y), theta = 75, fill = 2, individual = FALSE) +
    theme_bw() +
    theme(axis.title = element_blank())
ggplot(df) +
    geom_sector(aes(x, y), theta = 75, fill = 2, individual = FALSE) +
    coord_fixed(ratio = 3 / 5) +
    theme_bw() +
    theme(axis.title = element_blank())
ggplot(df) +
    geom_sector(aes(x, y), theta = 75, fill = 2, individual = TRUE) +
    theme_bw() +
    theme(axis.title = element_blank())

# theta
ggplot(df) +
    ## type = "percent", theta = 0-100
    geom_sector(
        aes(y, x, theta = value * 100),
        type = "percent",
        color = "blue",
        individual = TRUE
    ) +
    ## type = "degree", theta = 0-360
    geom_sector(
        aes(y, x, theta = value * 360),
        type = "degree",
        color = "red",
        alpha = 0.5,
        individual = TRUE
    ) +
    theme_bw() +
    theme(axis.title = element_blank())

# r
ggplot(df) +
    geom_sector(
        aes(y, x, theta = value * 100),
        r = rep(c(0.15, 0.3, 0.45), 5),
        fill = 2,
        individual = TRUE
    ) +
    theme_bw() +
    theme(axis.title = element_blank())

# start
ggplot(df) +
    geom_sector(
        aes(y, x, theta = value * 100),
        start = rep(c(60, 40, 20), 5),
        fill = 2,
        individual = TRUE
    ) +
    theme_bw() +
    theme(axis.title = element_blank())

# r_start
ggplot(df) +
    geom_sector(
        aes(y, x, theta = value * 100),
        r_start = rep(c(0.15, 0.25, 0.35), 5),
        fill = 2,
        individual = TRUE
    ) +
    theme_bw() +
    theme(axis.title = element_blank())
}

}
\keyword{datasets}
