% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%||\%}
\alias{html_pretty}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{prettydoc}{\code{\link[prettydoc]{html_pretty}}}

  \item{rlang}{\code{\link[rlang:op-null-default]{\%||\%}}}
}}

